chore: slight refactor of endpoint apply observation method
